Carrito actualizado.

+++ app/src/main/java/com/example/pressbeauty/model/CarritoUI.kt
@@ -4,5 +4,5 @@ data class CarritoUI(
     val idCarrito : String,
     val idUsuario : String,
     val productos: List<DetalleCarritoUI>,
-    val total : Float
+    val total : Int
 )
